Refactor: Rework armature creation logic in model separation
This change reorders the separation process by creating the new template
model (which becomes .001) before separating the bones from the source
model (which becomes .002).

The separated bones are then joined into the template armature,
and the temporary (.002) armature's data-block is immediately
purged. This results in a more logical workflow and prevents
lingering orphan data from the start.